Move rd_file_view_struct to header
